+++ internal/service/l3/investment.service.go
@@ -3,7 +3,6 @@ package l3_service
 import (
 	"context"
 	"database/sql"
-	"factorbacktest/internal"
 	"factorbacktest/internal/db/models/postgres/public/model"
 	"factorbacktest/internal/db/models/postgres/public/table"
 	"factorbacktest/internal/domain"
@@ -29,48 +28,6 @@ type InvestmentService interface {
 	Add(ctx context.Context, userAccountID uuid.UUID, savedStrategyID uuid.UUID, amount int) error
 	Reconcile(ctx context.Context) error
 	Rebalance(ctx context.Context) error
-}
-
-func AggregateAndFormatTrades(trades []*domain.ProposedTrade) []*domain.ProposedTrade {
-	// Map to hold aggregated trades by symbol
-	aggregatedTrades := make(map[string]*domain.ProposedTrade)
-
-	// Aggregate trades by symbol
-	for _, trade := range trades {
-		if existingTrade, exists := aggregatedTrades[trade.Symbol]; exists {
-			// Update the existing trade quantity
-			existingTrade.ExactQuantity = existingTrade.ExactQuantity.Add(trade.ExactQuantity)
-			aggregatedTrades[trade.Symbol] = existingTrade
-		} else {
-			// Add a new trade to the map
-			aggregatedTrades[trade.Symbol] = trade
-		}
-	}
-
-	// Create a slice to hold the formatted trades
-	var result []*domain.ProposedTrade
-	for _, trade := range aggregatedTrades {
-		if !trade.ExactQuantity.IsZero() {
-			result = append(result, trade)
-		}
-	}
-
-	// we could round all trades up to $1 but
-	// if they have tons of little trades, that
-	// could get expensive
-	// round all buy orders to $1
-	// TODO - i think we should use market value
-	// and figure out whether to round up or down
-	// also since price is stale, it could be just under $1
-	// also we need to ledger these somewhere, as excess that
-	// I own
-	for _, t := range trades {
-		if t.ExactQuantity.GreaterThan(decimal.Zero) && t.ExactQuantity.Mul(t.ExpectedPrice).LessThan(decimal.NewFromInt(1)) {
-			t.ExactQuantity = (decimal.NewFromInt(2).Div(t.ExpectedPrice))
-		}
-	}
-
-	return result
 }
 
 type investmentServiceHandler struct {
@@ -119,37 +76,6 @@ func NewInvestmentService(
 		AlpacaRepository:          alpacaRepository,
 		TradingService:            tradeService,
 	}
-}
-
-func (h investmentServiceHandler) listForRebalance() ([]model.Investment, error) {
-	investments, err := h.InvestmentRepository.List(repository.StrategyInvestmentListFilter{})
-	if err != nil {
-		return nil, err
-	}
-
-	investmentsToRebalance := []model.Investment{}
-	for _, investment := range investments {
-		tradeOrders, err := h.InvestmentTradeRepository.List(nil, repository.InvestmentTradeListFilter{
-			InvestmentID: &investment.InvestmentID,
-		})
-		if err != nil {
-			return nil, err
-		}
-		pendingInvestmentTradeID := uuid.Nil
-		for _, t := range tradeOrders {
-			if *t.Status == model.TradeOrderStatus_Pending {
-				pendingInvestmentTradeID = *t.InvestmentTradeID
-			}
-		}
-
-		if pendingInvestmentTradeID == uuid.Nil {
-			investmentsToRebalance = append(investmentsToRebalance, investment)
-		} else {
-			logger.Info("skipping rebalancing investment id %s: has pending investment trade %s\n", investment.InvestmentID, pendingInvestmentTradeID)
-		}
-	}
-
-	return investmentsToRebalance, nil
 }
 
 func (h investmentServiceHandler) Add(ctx context.Context, userAccountID uuid.UUID, savedStrategyID uuid.UUID, amount int) error {
@@ -220,85 +146,38 @@ func (h investmentServiceHandler) Add(ctx context.Context, userAccountID uuid.UU
 	return nil
 }
 
-type ComputeTargetPortfolioInput struct {
-	PriceMap         map[string]decimal.Decimal
-	Date             time.Time
-	PortfolioValue   decimal.Decimal
-	FactorScores     map[string]*float64
-	TargetNumTickers int
-	TickerIDMap      map[string]uuid.UUID
-}
-
-type ComputeTargetPortfolioResponse struct {
-	TargetPortfolio *domain.Portfolio
-	AssetWeights    map[string]float64
-	FactorScores    map[string]float64
-}
-
-// Computes what the portfolio should hold on a given day, given the
-// strategy (equation and universe) and value of current holdings
-// TODO - find a better place for this function
-func ComputeTargetPortfolio(in ComputeTargetPortfolioInput) (*ComputeTargetPortfolioResponse, error) {
-	if in.PortfolioValue.LessThan(decimal.NewFromFloat(0.001)) {
-		return nil, fmt.Errorf("cannot compute target portfolio with value %s", in.PortfolioValue.String())
-	}
-	if in.TargetNumTickers < 3 {
-		return nil, fmt.Errorf("insufficient tickers: at least 3 target tickers required, got %d", in.TargetNumTickers)
-	}
-
-	computeTargetInput := internal.CalculateTargetAssetWeightsInput{
-		Date:                 in.Date,
-		FactorScoresBySymbol: in.FactorScores,
-		NumTickers:           in.TargetNumTickers,
-	}
-	newWeights, err := internal.CalculateTargetAssetWeights(computeTargetInput)
+// listForRebalance retrieves all investments that should be
+// rebalanced right now
+// todo - fix so that it looks at rebalance interval
+func (h investmentServiceHandler) listForRebalance() ([]model.Investment, error) {
+	investments, err := h.InvestmentRepository.List(repository.StrategyInvestmentListFilter{})
 	if err != nil {
-		return nil, fmt.Errorf("failed to calculate target asset weights: %w", err)
+		return nil, err
 	}
 
-	// this is where the assumption that target portfolio will not hold
-	// cash comes from - the field is just not populated
-	targetPortfolio := domain.NewPortfolio()
-
-	// convert weights into quantities
-	for symbol, weight := range newWeights {
-		price, ok := in.PriceMap[symbol]
-		if !ok {
-			return nil, fmt.Errorf("priceMap does not have %s", symbol)
+	investmentsToRebalance := []model.Investment{}
+	for _, investment := range investments {
+		tradeOrders, err := h.InvestmentTradeRepository.List(nil, repository.InvestmentTradeListFilter{
+			InvestmentID: &investment.InvestmentID,
+		})
+		if err != nil {
+			return nil, err
 		}
-
-		// key line - determines how much new amount of symbol should be
-		// i want to round this to something so that we can generate results
-		// deterministically.
-
-		dollarsOfSymbol := in.PortfolioValue.Mul(decimal.NewFromFloat(weight)).Round(3)
-		quantity := dollarsOfSymbol.Div(price)
-
-		tickerID := uuid.Nil
-		if in.TickerIDMap != nil {
-			if id, ok := in.TickerIDMap[symbol]; ok {
-				tickerID = id
+		pendingInvestmentTradeID := uuid.Nil
+		for _, t := range tradeOrders {
+			if *t.Status == model.TradeOrderStatus_Pending {
+				pendingInvestmentTradeID = *t.InvestmentTradeID
 			}
 		}
 
-		targetPortfolio.Positions[symbol] = &domain.Position{
-			Symbol:        symbol,
-			ExactQuantity: quantity,
-			TickerID:      tickerID,
-			// if we want to switch to $ instead, add here
+		if pendingInvestmentTradeID == uuid.Nil {
+			investmentsToRebalance = append(investmentsToRebalance, investment)
+		} else {
+			logger.Info("skipping rebalancing investment id %s: has pending investment trade %s\n", investment.InvestmentID, pendingInvestmentTradeID)
 		}
 	}
 
-	selectedAssetFactorScores := map[string]float64{}
-	for _, asset := range targetPortfolio.Positions {
-		selectedAssetFactorScores[asset.Symbol] = *in.FactorScores[asset.Symbol]
-	}
-
-	return &ComputeTargetPortfolioResponse{
-		TargetPortfolio: targetPortfolio,
-		AssetWeights:    newWeights,
-		FactorScores:    selectedAssetFactorScores,
-	}, nil
+	return investmentsToRebalance, nil
 }
 
 func (h investmentServiceHandler) getTargetPortfolio(
@@ -705,7 +584,7 @@ func (h investmentServiceHandler) Rebalance(ctx context.Context) error {
 	// until we have some fancier math for reconciling completed trades,
 	// treat any failure here as fatal
 	// TODO - improve reconciliation + partial trade completion
-	executedTrades, tradeExecutionErr := h.aggregateAndExecuteTradeOrders(proposedTrades, rebalancerRun.RebalancerRunID)
+	executedTrades, tradeExecutionErr := h.TradingService.ExecuteBlock(proposedTrades, rebalancerRun.RebalancerRunID)
 
 	updateInvesmtentTradeErrors := []error{}
 	for _, tradeOrder := range executedTrades {
@@ -768,10 +647,4 @@ func proposedTradesToInvestmentTradeModels(trades []*domain.ProposedTrade, inves
 		})
 	}
 	return out
-}
-
-func (h investmentServiceHandler) aggregateAndExecuteTradeOrders(proposedTrades []*domain.ProposedTrade, rebalancerRunID uuid.UUID) ([]model.TradeOrder, error) {
-	aggregatedTrades := AggregateAndFormatTrades(proposedTrades)
-
-	return h.TradingService.ExecuteBlock(aggregatedTrades, rebalancerRunID)
 }
